apps/api/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0
httpx>=0.27.0

# Code Quality
black>=24.0.0
ruff>=0.4.0
mypy>=1.10.0

# Utilities
python-dotenv>=1.0.0
boto3>=1.34.0
requests>=2.31.0
Pillow>=10.0.0
